(cli) move ed25519 and ripemd160 to common #50
