+++ common/src/main/scala/pravda/common/contrib/ripemd160.java
@@ -1,4 +1,4 @@
-package pravda.node.contrib;
+package pravda.common.contrib;
 
 /*
  * Bitcoin cryptography library
